// src/app/models/models.ts
export interface MenuItems{
    path:string,
    menu:string
}
export interface PeriodicElement {
    name: string;
    position: number;
    weight: number;
    symbol: string;
  }
  
export interface FeedbackElement {
    uid: string;
    username: string;
    mobno: string;
    feedback: string;
    date:string;
    rootcause:string;
    id:number;
  }
  